Refactor: Change TPS546_get_temperature return type from int to float

+++ main/power/TPS546.h
@@ -175,7 +175,7 @@ void TPS546_read_mfr_info(uint8_t *);
 void TPS546_write_entire_config(void);
 int TPS546_get_frequency(void);
 void TPS546_set_frequency(int);
-int TPS546_get_temperature(void);
+float TPS546_get_temperature(void);
 float TPS546_get_vin(void);
 float TPS546_get_iout(void);
 float TPS546_get_vout(void);
